Cleanup

+++ setup.py
@@ -125,9 +125,14 @@ class build_ext(_build_ext):
         copy_file(
             os.path.join(cwd, "build", "src", "libzuspec-parser.so"),
             os.path.join(package_dir, "libzuspec-parser.so"))
-        copy_file(
-            os.path.join(cwd, "build", "antlr4", "lib64", "libantlr4-runtime.so"),
-            os.path.join(package_dir, "libantlr4-runtime.so"))
+        if os.path.isfile(os.path.join(cwd, "build", "antlr4", "lib64", "libantlr4-runtimetime.so")):
+            copy_file(
+                os.path.join(cwd, "build", "antlr4", "lib64", "libantlr4-runtime.so"),
+                os.path.join(package_dir, "libantlr4-runtime.so"))
+        else:
+            copy_file(
+                os.path.join(cwd, "build", "antlr4", "lib", "libantlr4-runtime.so"),
+                os.path.join(package_dir, "libantlr4-runtime.so"))
                 
         dest_filename = os.path.join(package_dir, filename)
